Modified the wget line in bin/build_setup.sh to get file from master
The file rUNSWift/bin/build_setup.sh originally pulled the GameController2017.zip
from the branch "Roclemir-patch-1", this was changed so the file is pulled
from the master branch instead.
Upon stable release to production code (as in, when this file is merged with
the production branch), this file should be updated so if pulls from the
production branch since it is the most stable branch.